Add a dashboard for capture sensors metrics
Starting from version 23.04, the capture sensors send their system and
processes metrics to the Prometheus instance of the Analytics deployment
they are resgistered to.

This commit add the Grafana dashboard to visualize those new metrics.

Note that this version of dashboard is limited to system and general
processes metrics only.
This is an attempt to have nearly the same observability than on PVX.
The extended sniffer metrics graphs will be add in ulterior version of
this dashboard.

References:
- CST - Building Grafana Dashboards to monitor Capture Sensors status:
  https://app.asana.com/0/1203491169098292/1204100287695974